[refactor] 브랜치 분리, 코드 다시 리팩토링.
